Add training code for Random Forest classifier of resiliency data

+++ second_attempt/include/Gamestate.hpp
@@ -42,7 +42,7 @@ class Gamestate {
 
   static float model_evaluate(Gamestate& g, bool smart_eval);
 
-  std::vector<float> resiliency_vector();
+  std::vector<double> resiliency_vector();
 
   std::pair<int, int> p1Pos;
   std::pair<int, int> p2Pos;
